Stripped the MeasuredGeometryProducer class from TotemRPGeometryESModule (to be reused in the new CTPPSGeometryESModule)
